Fix Compiler Warnings

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_VisualRenderer.cpp
@@ -327,9 +327,9 @@ void ERS_CLASS_VisualRenderer::UpdateViewport(int Index, ERS_CLASS_SceneManager*
 
         for (unsigned int x = 0; x < Model->AttachedScriptIndexes_.size(); x++) {
         
-            unsigned int ScriptIndex = Model->AttachedScriptIndexes_[x];
+            int ScriptIndex = Model->AttachedScriptIndexes_[x];
         
-            if (ScriptIndex > -1 && ScriptIndex < ProjectUtils_->ProjectManager_->Project_.Scripts.size()) {
+            if (ScriptIndex > -1 && (int)ScriptIndex < ProjectUtils_->ProjectManager_->Project_.Scripts.size()) {
                 NewIndexes.push_back(ScriptIndex);
             }
